+++ src/react/components/autocompleter/stories/variations.js
@@ -1,8 +1,9 @@
 const NOOP = require('../../helpers/noop');
+import { fixedWidthDropdownSizes } from '../../dropdown/private/constants';
 
 const storiesWithVariationsKindName = 'Instances/XUIAutocompleter';
 
-const dropdownSizes = ['small', 'medium', 'large', 'xlarge'];
+
 const inputSizes = ['small', 'medium'];
 
 const variations = [
@@ -44,12 +45,6 @@ const variations = [
 		storyKind: storiesWithVariationsKindName,
 		storyTitle: `shows a ${inputSize} input`,
 	})),
-	...dropdownSizes.map(dropdownSize => ({
-		storyKind: storiesWithVariationsKindName,
-		storyTitle: `shows a ${dropdownSize} dropdown`,
-		openDrawer: true,
-		dropdownSize,
-	})),
 	{
 		storyKind: storiesWithVariationsKindName,
 		storyTitle: 'shows a small picklist',
@@ -88,9 +83,18 @@ const variations = [
 	}
 ];
 
+Object.keys(fixedWidthDropdownSizes).map(dropdownSize =>
+	variations.push({
+		storyKind: storiesWithVariationsKindName,
+		storyTitle: `shows a ${dropdownSize} dropdown`,
+		openDrawer: true,
+		dropdownSize,
+	})
+);
+
 module.exports = {
 	variations,
 	NOOP,
 	storiesWithVariationsKindName,
-	dropdownSizes,
+	fixedWidthDropdownSizes,
 };