+++ src/react/components/dropdown/DropDown.js
@@ -5,6 +5,7 @@ import DropDownLayout from './DropDownLayout';
 import DropDownPanel from './DropDownPanel';
 import { lockScroll, unlockScroll } from '../helpers/lockScroll';
 import { ns } from '../helpers/xuiClassNamespace';
+import { fixedWidthDropdownSizes } from './private/constants';
 
 /**
  * Wrapper for all content which will go inside of a dropdown.  It ensures the correct
@@ -235,9 +236,9 @@ DropDown.propTypes = {
 	/** Whether or not this component is hidden. */
 	isHidden: PropTypes.bool,
 
-	/** Applies the correct XUI class based on the chose size. Default will
-	 * fits to children's width. */
-	size: PropTypes.oneOf(['small', 'medium', 'large', 'xlarge']),
+	/** Applies the correct XUI class based on the chosen size. Default will
+	 * fit to children's width. */
+	size: PropTypes.oneOf(Object.keys(fixedWidthDropdownSizes)),
 
 	/** An array of keydown keycodes to be ignored from dropdown behaviour. */
 	ignoreKeyboardEvents: PropTypes.array,
@@ -257,7 +258,7 @@ DropDown.propTypes = {
 	/** Whether or not the dropdown should take focus and handle keyboard events automatically */
 	hasKeyboardEvents: PropTypes.bool,
 
-	/** Callback for adding additional onKeyPress funcitonality */
+	/** Callback for adding additional onKeyPress functionality */
 	onKeyDown: PropTypes.func,
 
 	/** Whether focus should be restricted to the dropdown while it's open. */
@@ -284,7 +285,7 @@ DropDown.propTypes = {
 	/** Force the desktop UI, even if the viewport is narrow enough for mobile. */
 	forceDesktop: PropTypes.bool,
 
-	/** Force wrapping Panel childrens in a StatefulPicklist  */
+	/** Force wrapping Panel children in a StatefulPicklist  */
 	forceStatefulPicklist: PropTypes.bool,
 
 	/** Class to apply to the body element of the dropdown */