+++ changes.md
@@ -79,6 +79,7 @@ We recommend running a bundle analyzer after upgrading (and regularly in general
 	- The new Pickitem prop `isInvalid` styles the item in an invalid state
 	- `NestedPicklistTrigger` also supports a `leftElement`
 	- In support of the more contentblock-style Autocompleter picklists, Pickitems also have new `headerElement` and `isMultiline` props.
+	- Pickitem also now supports onBlur and onFocus handlers
 - ContentBlockItem has new `description` and `tagPosition` props. Description is styled similarly to the prior `secondaryHeading` content (which has now become more prominent), and `tagPosition` allows you to specify where tags will be placed. `pinnedValue`, `href`, `primaryHeading`, and `secondaryHeading` now only accept strings. The `tag` property has been changed to accept multiple tags, and is now labeled `tags`, accordingly.
 - Loader size options have been switched from 'large', 'standard', and 'small' to be consistent with other component sizes. New size options are 'medium', 'small', and 'xsmall', and will be converted automatically if using the codemod. The largest size, 'medium' is now the default.
 - `SelectBox`
@@ -153,6 +154,7 @@ Utility classes with breakpoint suffixes have been updated to reflect the new br
 
 ### Component props
 
+- Dropdown has new values for `size` prop; `xsmall`, `small`, `medium` and `large`.
 - Pill has a new `isLimitedWidth` prop to replace the `isMaxContentWidth` prop. The default for pills is now to fit their content, and you can apply `isLimitedWidth` to cap them at 200px. This prop change will be handled automatically by the upgrade codemod for existing Pills in your app.
   - Pill `size` prop now has a default of `medium`, instead of `standard`. This size is visually the same as XUI 14 aside from the font size changes. Other available values are `small` and `xsmall`.
 - XUIIcon, when `isBoxed` is true, now adds a class called `xui-iconwrapper-medium` by default. Use the `size` prop to alter this to `large` or `xlarge` at your discretion.
@@ -175,3 +177,4 @@ Utility classes with breakpoint suffixes have been updated to reflect the new br
 
 - Modal widths have changed: small is 300px (was 340); medium is 400px (was 460); large is 600px (was 600) and xlarge is 800px (was 860).
 - In addition to renaming the utility classes that handle viewport-size-specific behaviour, the behaviour itself has changed, in some cases. `-medium` was previously affecting elements when the viewport was *between* 520px and 940px. The new `-small-up` replacement classes will affect elements when the viewport is 600px or greater, unless overridden with a `-large-up` class, which will take over at 1000px. This is in line with a mobile-first approach. In many cases, this will mean that developers need fewer classes to get the desired behaviour.
+- Along with new sizes being introduced for dropdown, the values of dropdown sizes have also changed: xsmall is 100px (xsmall previously didn't exist); small is 200px (was 140px); medium is 300px (was 220px) and large is 400px (was 300px). xlarge has been removed.